Crate trillium_logger

Welcome to the trillium logger!

Re-exports

pub use crate::formatters::apache_combined;

pub use crate::formatters::apache_common;

pub use crate::formatters::dev_formatter;

Modules

formatters

Components with which common log formats can be constructed

Structs

Logger

The trillium handler for this crate, and the core type

Enums

ColorMode

A configuration option that determines if format will be colorful.

Target

Specifies where the logger output should be sent

Traits

LogFormatter

The interface to format a &Conn as a Display-able output

Functions

logger

Convenience alias for Logger::new